Move auth under server/apis
